In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,13 +9,13 @@
 	int main(void)
 	{
 		int count;
-		int fib1 = 0, fib2 = 1, sum;
+	unsigned long fib1 = 0, fib2 = 1, sum;
 
 
 		for (count = 0; count < 50; count++)
 		{
 			sum = fib1 + fib2;
-			printf("%d", sum);
+			printf("%lu", sum);
 
 
 			fib1 = fib2;
